pyproject.toml
[project]
name = "asu-app"
version = "0.1.0"
description = "Buddy4U - Future starts now!"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.9"

dependencies = [
  "streamlit",
  "pandas",
  "matplotlib",
  "Pillow",
  "streamlit-extras"
]
